#21 fix: 검색 텍스트필드 뷰 따로 빼기
