+++ ShallWe/ShallWe/Presentation/Main/Search/ViewControllers/SearchViewController.swift
@@ -24,12 +24,7 @@ final class SearchViewController: BaseViewController {
     // MARK: - UI Components
     
     private let navigationBar = CustomNavigationBar()
-    private let searchView = UIView()
-    private let searchTextField = UITextField()
-    private let searchIconView = UIImageView()
-    private lazy var deleteTextButton = UIButton()
-    private lazy var cancelButton = UIButton()
-    private let underLineView = UIView()
+    private let searchView = SearchTextFieldView()
     private lazy var resultLabel = UILabel()
     private lazy var searchResultView = SearchResultView()
     private lazy var recentSearchView = RecentSearchView()
@@ -53,7 +48,7 @@ final class SearchViewController: BaseViewController {
             })
             .disposed(by: disposeBag)
         
-        cancelButton.rx.tap
+        searchView.cancelButton.rx.tap
             .subscribe(onNext: { [weak self] in
                 self?.popToHomeVC()
             })
@@ -71,35 +66,6 @@ final class SearchViewController: BaseViewController {
             $0.isBackButtonIncluded = true
         }
         
-        searchView.do {
-            $0.backgroundColor = .bg2
-            $0.makeCornerRound(radius: 37 / 2)
-        }
-        
-        searchIconView.do {
-            $0.image = ImageLiterals.Icon.search
-        }
-        
-        searchTextField.do {
-            $0.font = .fontGuide(.R00_14)
-            $0.textColor = .black
-        }
-        
-        deleteTextButton.do {
-            $0.setImage(ImageLiterals.Icon.delete, for: .normal)
-            $0.isHidden = true
-        }
-        
-        cancelButton.do {
-            $0.setTitle("취소", for: .normal)
-            $0.titleLabel?.font = .fontGuide(.R00_14)
-            $0.setTitleColor(.line1, for: .normal)
-        }
-        
-        underLineView.do {
-            $0.backgroundColor = .bg2
-        }
-        
         resultLabel.do {
             $0.font = .fontGuide(.SB00_14)
             $0.textColor = .gray4
@@ -114,9 +80,8 @@ final class SearchViewController: BaseViewController {
     
     override func setLayout() {
         
-        self.view.addSubviews(navigationBar, searchView, underLineView,
+        self.view.addSubviews(navigationBar, searchView,
                               resultLabel, searchResultView, recentSearchView)
-        searchView.addSubviews(searchIconView, searchTextField, deleteTextButton, cancelButton)
         
         navigationBar.snp.makeConstraints {
             $0.top.equalTo(view.safeAreaLayoutGuide)
@@ -125,52 +90,23 @@ final class SearchViewController: BaseViewController {
         }
         
         searchView.snp.makeConstraints {
-            $0.top.equalTo(navigationBar.snp.bottom).offset(13)
-            $0.horizontalEdges.equalToSuperview().inset(16)
-            $0.height.equalTo(37)
-        }
-        
-        searchIconView.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.leading.equalToSuperview().inset(14)
-        }
-        
-        searchTextField.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.leading.equalTo(searchIconView.snp.trailing).offset(16)
-            $0.width.equalTo(147)
-            $0.height.equalTo(28)
-        }
-        
-        deleteTextButton.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.trailing.equalToSuperview().inset(13)
-        }
-        
-        cancelButton.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.trailing.equalToSuperview().inset(17)
-            $0.size.equalTo(28)
-        }
-        
-        underLineView.snp.makeConstraints {
-            $0.top.equalTo(searchView.snp.bottom).offset(25)
-            $0.horizontalEdges.equalToSuperview().inset(24)
-            $0.height.equalTo(1)
+            $0.top.equalTo(navigationBar.snp.bottom)
+            $0.horizontalEdges.equalToSuperview()
+            $0.height.equalTo(63)
         }
         
         resultLabel.snp.makeConstraints {
-            $0.top.equalTo(underLineView.snp.bottom).offset(109)
+            $0.top.equalTo(searchView.snp.bottom).offset(109)
             $0.centerX.equalToSuperview()
         }
         
         searchResultView.snp.makeConstraints {
-            $0.top.equalTo(underLineView.snp.bottom)
+            $0.top.equalTo(searchView.snp.bottom)
             $0.horizontalEdges.bottom.equalToSuperview()
         }
         
         recentSearchView.snp.makeConstraints {
-            $0.top.equalTo(underLineView.snp.bottom)
+            $0.top.equalTo(searchView.snp.bottom).offset(12)
             $0.horizontalEdges.bottom.equalToSuperview()
         }
     }
